Update patch set 1
Patch Set 1: Code-Review-1

(40 comments)

Hi Uli,

This looks like a good start, I added a few comments and suggestions in line. A few references are not good, I added the correct ones in comments as well.

Also please try to keep the line length under the recommended, that helps with the readability of the review.

Thanks and Best Regards,
Ildiko

Patch-set: 1
Label: Code-Review=-1