Update patch set 2
Patch Set 2:

Hi Gerald,  
Yes I thiknk alignment on titles would be best.  That we can probably take as a Jira ticket on stable after release as some of the repo's have already release tagged.  (plus coordinating that takes a lot of time)
The platformoverview doc was pretty last minute and we did not manage to get everything in.  Next release we should look to improve.

Patch-set: 2
Label: Verified=0