Update patch set 1
Patch Set 1: Code-Review-1

Looking at this change, the URL for the rendered document will be .../install-config/index.html

I think we should discuss the directory name as it will essentially be the URI for the document (the name) in Colorado.  Having a "-" is not ideal and we could potentially choose a more descriptive name.

Patch-set: 1
Reviewer: Gerrit User 163 <163@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1