improve config
A developer was asked to copy all contents in the base config file in
order to set all parameters properly, even if the change is partial.

This patch improves this config by setting all default parameters based
on users' custom config file, so that they can write parameters just
which they want to change.

This patch also fixes bug in build-composite.sh .

Change-Id: Ide459c8e9f416b29b114a4bf2a432e8eae1f841e
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>