Creating a template for feature projects to produce userguide and scenario description documents
This template will be pushed to features projects in Colorado to help them create the right
documentation for the Colorado release.  The documentation will be feature and scenario specific
at a level where that makes sense while allowing common activities like instalation to be
handled in documents for the installation tools.

Change-Id: I649c583233ab8d8f8c3ebf7ddf8ae539dae15c8a
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>