Update patch set 1
Patch Set 1: Code-Review+1

What I would include in the scenario description for the specific case of SDNVPN is how to set up the scenario manually through the Fuel GUI, in case someone wants to have an easy way of customizing the scenario. But I guess that does not need to be reflected in the template.

Patch-set: 1
Reviewer: Gerrit User 321 <321@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1