Update patch set 3
Patch Set 3: Code-Review+2

We need to start somewhere, this is in accordance to what we have discussed in the docs team. And we will of course be open for improvements, but we need to move forward. Chris please merge this whenever you feel ready.

Patch-set: 3
Reviewer: Gerrit User 2409 <2409@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2