WIP: Created 2 introduction documents for user guide and installation
Work on the contant in both installation and user guide

Started to add artifacts in the installation instruction  and user guide

This is a first draft of the installation.introduction.rst and userguide.introduction.rst.
Some of the information has been reused from the Brahmaputra docs with  modifications.
Please comment and add as needed/wanted.

Change-Id: I0c6d79aac7d1152ec894986dda251ad2d3c7048e
Signed-off-by: Sofia Wallin <sofia.wallin@ericsson.com>