Update patch set 7
Patch Set 7:

I think that many projects do not publish requirements because we do not provide a place and process for doing so.  
Let's do this, I'll migrate all project documentation together in this rendering.  We can then #action sofia to provide clear guidance for Danube on how docs will be published.

Patch-set: 7
Label: Verified=0