Update patch set 12
Patch Set 12:

This one is intended to remain on master and can be our template for document presentation.  People can add patches during Danube as they complete new documents.
For Colorado we will cherry pick this then adjust the URL's.  We can also label it there if we like.

Patch-set: 12