Update patch set 1
Patch Set 1:

> are you saying jenkins jobs? we don't use jenkins job chain for
 > opnfv docs so far. and, I'm not a fan of jenkins job chaining as it
 > could consume one job slot in CI pipeline.

Not exactly. An example for propagation error is that when the parent folder is missing, all operation to the underlying files will fail.

Anyway, I shall take your advice since analysing root cause is overkilled for this script. We may leave it to the user to inspect the error messages.

Patch-set: 1
Label: Verified=0