Update patch set 1
Patch Set 1: Code-Review-1

I think I need to understand how you are thinking here. If this is how we need to do this it need to be structured i some way. And if this is docs/testing/ there are several projects included here that are not a test project.

Patch-set: 1
Reviewer: Gerrit User 2409 <2409@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1