Update patch set 2
Patch Set 2: Code-Review+1

Looks OK with the method seeming to work for importing.  Let's set aside an hour to discuss on the next call and ensure we are all on the same page.
I'd also like to see us arrive at a preferred format/layout for this as well.

Patch-set: 2
Label: Code-Review=+1