Update patch set 3
Patch Set 3:

(2 comments)

Great, but we need to add a template for the configuration guide as well. I think that the SFC config guide written for Colorado can be used as a template for Danube. And to have the release name in the titles feels a bit redundant, but maybe someone else has another opinion.

Patch-set: 3