Update patch set 3
Patch Set 3:

> please amend this patch to include the relevant documentation
 > changes

There was a call with ODLDocs yesterday and a proposal to keep common merge and verify jobs. If that proposal is to be taken up, we'll need to keep the build job same as ODL. Can we hold on to this for another day or so till we have a confirmation from Sofia/ Fatih on what approach is to be followed?

Patch-set: 3