New Tox build script
To keep the doc jobs for OPNFV & ODL as similar as possible, we're
using a similar tox build script as used in ODLDocs.  The new merge
job is tested to run the same tox build script.  This build script
is to be run as `tox -edocs`. Everything under docs/ shall be built.

Change-Id: Ic4fab5e431e6937a1dbbeca2b1ff339dd092b688
Signed-off-by: ShubhamRathi <shubhamiiitbackup@gmail.com>