Update patch set 3
Patch Set 3:

> Even if we aim to work as cross collaborative as possible we should
 > first off all make sure that all builds runs correctly for OPNFV.

I have made the a new build script with Tox ( similar to what is used in ODLDocs). Intent of keeping the build script same as ODL is to keep the option of collaboration open in future. Plus, since they already have a robust documentation system in place - it makes perfect sense to follow their precedent. If its agreeable, I propose that we abandon this patch in favor of the new tox build script.

Patch-set: 3