Update patch set 5
Patch Set 5:

> > (5 comments)
 > 
 > I mean that the links to the subpages needs to be removed from the
 > wiki, as we discussed. But first we need to make sure that the 'how
 > to add your documentation' section is added on the main page.

Right. I'll remove it from the wiki tonight. I've added the include documentation part. 
I was wondering if its a good idea to add to this documentation on how to build the documentation locally? Tox script is for consolidated build which could also be used but I guess its nicer to go with the inbuilt sphinx-build command and explain how the build works in 3 simple steps. What do you think?

Patch-set: 5