Update patch set 1
Patch Set 1:

> > (10 comments)
 > 
 > Ok, so what we are saying here http://artifacts.opnfv.org/opnfvdocs/docs/how-to-use-docs/documentation-example.html
 > is not correct then?
 > It needs to be?
 > Title
 > -----

No. What we are saying there is correct. Even in the artifact link you've sent, there's distinction in the markup for chapters and sections (http://artifacts.opnfv.org/opnfvdocs/docs/how-to-use-docs/documentation-example.html#source-files). The chapter level markup is picked up by the toctree
Rule of thumb is that in the whole page, there has to be be only 1 chapter markup (heading)

Patch-set: 1
Label: Code-Review=0