Applying OPNFV style also to stable/danube branch
This patchset applies the OPNFV-like style to the documentation in the
stable/danube branch.

There is a style-sheet in the docs repo which is supposed to customize
the default style, but which was actually never applied. This patchset
enables custom overrides of the default style.

As one of the custom style changes, this patchset changes the color of
hyperlinks to the same green used on www.opnfv.org and increases the
fontsize to 16px. It also alignes the font families used by just sticking
to Helvetica.

Change-Id: I9736ec1109ba272104f861758fd7215124dc631b
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>