Enabling custom style overrides and introducing green links
There is a style-sheet in the docs repo which is supposed to customize
the default style, but which was actually never applied. This patchset
enables custom overrides of the default style.

As one of the custom style changes, this patchset changes the color of
hyperlinks to the same green used on www.opnfv.org.

Change-Id: I3c8a59e7f0beaf0150915941730f68071a40330a
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>
(cherry picked from commit 9b153bc8882ad8b4bd52ff7a412118dd48c06d63)