Remove infra repo from submodules
Having infra in submodules result in generation of duplicated
documentation such as for Pharos. As discussed, it is probably
logical to keep source of the project documentation in corresponding
project repos so they can be developed alongside the actual software
so they can be reviewed and approved by the project committers.

Change-Id: Ifda50d25d9a1b1e9bcf373713dc759f140fb046a
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>