Update patch set 2
Patch Set 2:

> (1 comment)

Since the container is only used for docs building. I would suggest to keep all the preparation in the Dockerfile, such as virtualenv(I'm also not sure whether it is necessary to use virtualenv in the docker), tox, and pip but use the same entry as:
tox -edocs

I we don't care about the run time, we can use the same.

Patch-set: 2
Label: Verified=0