Update OPNFV Sphinx Theme
When the theme was initially created I was under the assumption the
directory the theme lives under needs to match the name. This does not
seem to be the case and in fact seems to cause issues with correctly
importing and using the theme.

Includes formatting updates to the README and linting.

Change-Id: I67c834bfe9dc99252159394ecbf136d9a264e0a8
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>