Intersphinx Link Documentation to release Branch
This is an example of how we can set the links for projects following
the stable branch process.

When opnfvdocs is branched for Gambia, the version/release will be updated to
'Gambia', and branch will be set to 'stable-gambia'.

With the intersphinx_mapping links referencing 'branch', the links will
be updated to match the change.

Change-Id: Icf25cfb3e36a5b8cb533abbf4b929781c4e5fa09
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>