Set Version and Release to Hunter
Updates the 'branch' config which is used for intersphinx linking, and
sets the version/release to Hunter.

Switches projects not in this release to link against their latest docs
as they don't have a 'stable-hunter' to reference.

Includes an update to '.gitreview' to ensure `git-review` submits a
patch against the correct branch.

Change-Id: Id2f46b8fd7f71eb3e46b7698896ab1d4059345f7
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>