Update patch set 6
Patch Set 6:

So this is the problem of not tracking what documentation the projects plan to deliver. 
VSPerf documentation is listed here, https://docs.opnfv.org/en/latest/testing/testing-user.html

Please also remember that it is also the projects responsibility to ensure that their documentation is listed on docs.opnfv.org

Airship and Kuberef are added to the config and user guide. But I can move the link for airship to the installation doc instead if that’s preferred. 


> Patch Set 6:
> 
> (1 comment)
> 
> So for AIRSHIP, there should be a reference to the Installation Guide in addition to the Release notes (already there).
> 
> VSPerf is an "old" project, It's hard to believe that they have no User Guide this time ??? 
> 
> there may be something small to fix here, to find all the project docs...

Patch-set: 6