Update patch set 7
Patch Set 7:

I'm not sure if I'm not reading this correct. The documentation for VSPERF and SampleVNF is available under "Testing Frameworks" and it has never been part of the Configuration and User guide. Moving this is of course ok but just to clarify from a structure point of view. 

I moved the Airship and Kuberef to the installation since both projects provide an installation guide. 

> Patch Set 6:
> 
> Thanks for checking the repos for Docs, Sofia.
> 
> Since VSPERF has combined their Config and User Guides into one doc, I would like to see the same link in both the Configuration and User Guide lists. The VSPerf Test Guide is also for Users, please Ref that one too under User Guides. 
> 
> I see that SampleVNF has a User Guide on the URL/page you shared, please add that on to the User Guides section.
> 
> THANKS!

Patch-set: 7