Add RTD Configuration file
Builds for the master branch broke on RTD broke with the move of the
requirements.txt file from etc/. Updating the path on RTD for
requirements fixes this, but at the cost of breaking all the historic
stable branch builds.

Using .readthedocs.yml the branch can be configured with it's own
location for requirements separate from the global configuration.

The setting to disable cloning of submodules is also included in this
configuration.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>