Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3: Code-Review-1
> 
> no pptx in tree.
> opnfvdocs must be fully rewritten...

Hi, thanks for the comment. Can you suggest an other way to store the editable files for our figures? Maybe we should other tools than PowerPoint, but that does not seem to feasible to many Anuket contributors. 

I agree that eventually we will need to rewrite opnfvdocs and at the same time I would like to move it to GitHub, but I would do it in a separate changeset/pr than this. With this I would like to focus on the landing page.

Patch-set: 3